Implement fuzzy search in rust
This was more difficult than I care to admit. I never quite realized how
much mutable state I rely on when doing development. Originally, I was
attempting a looping solution that modifies state-in-place. Eventually,
something clicked and I went the route of recursive function calls.

Definitely not as elegant as it could be, but it was a lot of fun to
work through.